Extend invoice to budget mapping by variant query
According to the Aleph documentation, Z601_TYPE "ENC" establishes a link between item and budget, not invoice and budget. The latter would be Z601_TYPE "INV". We are missing a lot of budgets for invoices  (20%) if we only consider Z601_TYPE "INV". After sampling the database for examples, it seems like mapping invoice to budget using Z601_TYPE "ENC" also yields correct results. For now, we consider the official documentation unclear about the mapping and use both "INV" and "ENV" Z601_TYPE for mapping the invoices to budgets.